+++ test/index.js
@@ -155,6 +155,66 @@ describe('Logging', function () {
 		log.setPrefix('TEST2');
 		logger.debug('prefix updated');
 	});
+
+	it('Can create table data', function () {
+		var test = {
+			one: {
+				id: 0,
+				name: 'testOne',
+				value: 1,
+				language: '日本語',
+			},
+			two: {
+				id: 1,
+				name: 'testTwo<2>',
+				language: 'こんにちは',
+				ident: '02'
+			},
+			three: {
+				id: 2,
+				name: 'testThree<3>',
+				value: 3,
+				language: 'ありがとうございます',
+				list: ['A', 'B', 'C'],
+				map: { c: 'c', d: 'd', e: [0,1,2,3,4,5], f: { a:'A',b:'B',c:'C'}, g: '{one:1, two:2, three:3}' }
+			},
+			four: {
+				id: 3,
+				name: 'testFour12345678901234567890123456789012345678901234567890123456789012345678901234567890',
+				value: 4,
+				language: 'Москве́',
+				map: { a: 'a', b: 'b' }
+			},
+			five: {
+				name: '魏民以夜光為怪石',
+				language: 'chinese'
+			},
+			six: {
+				name: 'korean',
+				value: -200,
+				language: '해빛',
+				list: [10, 20, 30]
+			},
+			seven: {
+				name: 'lorem ipsam',
+				value: '، ان بعد, ثم عدم وقامت الآلاف للمجهود',
+				language: 'arabic',
+				list: [5, 6],
+				map: { a: 500, b: 'd' },
+				ident: '0056'
+			}
+		};
+		var list = [1, 2, 3, 4, 5, 10, 123456];
+		var table1 = logger.createTable(test);
+		var table2 = logger.createTable(list);
+		logger.verbose(table1, table2);
+		logger.debug(table1, table2);
+		logger.trace(table1, table2);
+		logger.info(table1, table2);
+		logger.warn(table1, table2);
+		logger.error(table1, table2);
+		logger.fatal(table1, table2);
+	});
 	
 	it('Can remove all log files from the test (' + process.cwd() + '/test/logs/*.log)', function (done) {
 		var exec = require('child_process').exec;